Move controller ownership helpers into chrome
Constraint: Utilities cleanup should move one ownership group per PR without compatibility shims when the import surface is manageable.\nRejected: Moving pid-manager and process guardian in the same PR | those touch launcher cleanup and orphan reaping, so they deserve a separate verification slice.\nRejected: Leaving controller-lock under utils | it models Chrome controller/profile ownership, not a cross-cutting leaf utility.\nConfidence: high\nScope-risk: moderate\nDirective: Put Chrome controller/profile ownership helpers under src/chrome; keep src/utils for domain-neutral leaf helpers.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/chrome/controller-lock.test.ts tests/chrome/duplicate-controller-diagnostics.test.ts tests/cli/doctor/runtime-diagnostics.test.ts tests/cli/doctor/checks/duplicate-controllers.test.ts tests/tools/connection-health.test.ts tests/transports/duplicate-controller-error-server.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
