+++ tests/transports/duplicate-controller-error-server.test.ts
@@ -3,7 +3,7 @@ import {
   DUPLICATE_CONTROLLER_ERROR_CODE,
   DuplicateControllerErrorServer,
 } from '../../src/transports/duplicate-controller-error-server';
-import { DuplicateControllerError, type ControllerLockMetadata } from '../../src/utils/controller-lock';
+import { DuplicateControllerError, type ControllerLockMetadata } from '../../src/chrome/controller-lock';
 
 function makeServer(): DuplicateControllerErrorServer {
   const owner: ControllerLockMetadata = {
